// src/http-courses/models/empregabilidadeCurriculoConquista.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * API Go
 * API de serviços para aplicativos da Prefeitura do Rio
 * OpenAPI spec version: 1.0
 */
import type { EmpregabilidadeTipoConquista } from './empregabilidadeTipoConquista'

export interface EmpregabilidadeCurriculoConquista {
  cpf?: string
  created_at?: string
  descricao?: string
  id?: string
  id_tipo_conquista?: string
  /** Relationships */
  tipo_conquista?: EmpregabilidadeTipoConquista
  titulo?: string
  updated_at?: string
}