// src/http-courses/models/empregabilidadeCurriculoExperiencia.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API Go
 * API de serviços para aplicativos da Prefeitura do Rio
 * OpenAPI spec version: 1.0
 */

export interface EmpregabilidadeCurriculoExperiencia {
  cargo?: string
  cpf?: string
  created_at?: string
  descricao_atividades?: string
  eh_trabalho_atual?: boolean
  empresa?: string
  experiencia_comprovada_ct?: boolean
  id?: string
  tempo_experiencia_meses?: number
  updated_at?: string
}